requirements.txt
matplotlib>=3.1.3
torch>=1.1.0
opencv-python==4.9.0.80
numpy>=1.18.1

Flask~=3.0.0
APIFlask~=2.1.0
altair~=5.2.0
pandas~=2.2.0
streamlit~=1.30.0
pillow~=10.2.0